refactor(user): create user package
