+++ src/test/java/woowacourse/acceptance/AuthAcceptanceTest.java
@@ -16,7 +16,7 @@ import org.springframework.http.MediaType;
 import woowacourse.auth.dto.TokenRequest;
 import woowacourse.auth.dto.TokenResponse;
 import woowacourse.setup.AcceptanceTest;
-import woowacourse.shoppingcart.dto.request.SignUpRequest;
+import woowacourse.user.dto.request.SignUpRequest;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayName("인증 관련 기능")
